requirements.txt
numpy 
pandas 
Pillow
transformers
diffusers
accelerate 
wandb
lpips